use Sass/scss for styles, gulp for livereload
CSS styles are now written in the scss variant of Sass and when the
gulp process is running changes to the scss file(s) will automatically
kick off a process that generates the actual CSS files.

The gulp process also is using the livereload plugin and any changes
to the html, css, javascript, or json files will cause the browser
to reload the page.

The project now uses node and the node-package-manager to setup the
local development environment.