add Toggle Advanced Tools button
Shows or hides the Scaling control and the Image Layer Histogram.

The button actions toggles presence of class "show" in all elements
with class "developer"

Using a MutationObserver to respond to changes in class attributes
when "showing" an element like the Image Layer Histogram that need to rerun the rendering operation.

Toggle Full Screen control now uses Bootstrap icons.

It can also be used on other elements like the main image.