main content can also be made fullscreen
Event handlers can only be attached to components after the html
of the component has been inserted into the dom. In order to make it
easier to keep the code for the event handler near the code that renders
the html the rendering code for component returns a callback function
to the top-level renderer where it is saved in an array. After html
rendering is complete these callback functions are run.

See: 

renderDev.developerToolsButton = (containerId, registerCallback)

This pattern is extended in the code that renders the full-screen
buttons. In this case the event handler in the component needs
access to lexically bound variables in the top level renderer so a
private function in the top level renderer that has access to the
lexically bound variables is passed as an argument when the event
hander callback is run.

See:

renderDev.fullScreenButton(id2, '#main-image-content', renderedCallbacks, optionalFunc)

in: ./modules/render/activity.js

and:

renderDev.fullScreenButton = (containerId, epandedElement, registerCallback, optionalFunc)

in: /modules/render/dev.js
renderDev.fullScreenButton