masterpiece linear/log layer histogram updating
linear/log layer histogram also updating when selecting new layer
in rgb and multi-wave pages

Under the hood;

The Page instance manages changes in the json image object so don't
clone it when making a CanvasImage instance. Instead create instance
variable references to objects in image and getters for reading
properties in the image object managed by the Page instance.