find-apollo change border color site, zoomrect
Change border color for Apollo landing site and "You are here" zoom rect
to bright pink when zoom scale matches scale of landing site image.

Currently the landing site image scale is based on the maxScale1to1
scaling.

TODO: add location of landing site and parameters for size and offset
of landing site image and render it from the data.