wip observations
Additional observation targets.

Each target has a start and end date for scheduling and is only
shown if current date is scheduled.

For development a tester can add a four character date to the url
and this parameter will be used for current month and day.

Example: render list of targets as if current date is October 31

/#menu/observation?now=1031